backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database - PostgreSQL
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Database - MongoDB
motor==3.3.2
pymongo==4.6.1

# Task Queue
celery==5.3.6
redis==5.0.1

# HTTP Client and Scraping
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.1.0
pandas==2.2.0

# AI - Google Gemini
google-generativeai>=0.8.0
requests==2.31.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Logging and Monitoring
loguru==0.7.2

# Authentication
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
email-validator>=2.1.0
